+++ plugins/source/azure/resources/services/postgresql/server_firewall_rules_mock_test.go
@@ -1,0 +1,29 @@
+package postgresql
+
+import (
+	"encoding/json"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/postgresql/armpostgresql"
+	"github.com/cloudquery/plugin-sdk/faker"
+	"github.com/gorilla/mux"
+	"net/http"
+)
+
+func createFirewallRules(router *mux.Router) error {
+	var item armpostgresql.FirewallRuleListResult
+	if err := faker.FakeObject(&item); err != nil {
+		return err
+	}
+
+	router.HandleFunc("/subscriptions/{subscriptionId}/resourceGroups/debug/providers/Microsoft.DBforPostgreSQL/servers/test string/firewallRules", func(w http.ResponseWriter, r *http.Request) {
+		b, err := json.Marshal(&item)
+		if err != nil {
+			http.Error(w, "unable to marshal request: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		if _, err := w.Write(b); err != nil {
+			http.Error(w, "failed to write", http.StatusBadRequest)
+			return
+		}
+	})
+	return nil
+}